refactor: merge file-ops.ts into file.ts and resolve circular dependency
- Consolidate basic file operations (ensureDir, readFileContent, writeFileContent, fileExists) into file.ts
- Remove circular dependency between file.ts and ignore.ts by:
  - Removing filterIgnoredFiles import from file.ts
  - Updating parser.ts to use both findFiles and filterIgnoredFiles separately
- Update all imports from file-ops.js to file.js
- Remove redundant file-ops.ts file
- All tests pass (360/360) after refactoring
