+++ src/utils/file.ts
@@ -1,26 +1,36 @@
-import { readdir, rm } from "node:fs/promises";
-import { join } from "node:path";
-import { ensureDir, fileExists, readFileContent, writeFileContent } from "./file-ops.js";
-import { filterIgnoredFiles } from "./ignore.js";
+import { mkdir, readdir, readFile, rm, stat, writeFile } from "node:fs/promises";
+import { dirname, join } from "node:path";
 
-export { ensureDir, fileExists, readFileContent, writeFileContent };
+export async function ensureDir(dirPath: string): Promise<void> {
+  try {
+    await stat(dirPath);
+  } catch {
+    await mkdir(dirPath, { recursive: true });
+  }
+}
 
-export async function findFiles(
-  dir: string,
-  extension: string = ".md",
-  ignorePatterns?: string[],
-): Promise<string[]> {
+export async function readFileContent(filepath: string): Promise<string> {
+  return readFile(filepath, "utf-8");
+}
+
+export async function writeFileContent(filepath: string, content: string): Promise<void> {
+  await ensureDir(dirname(filepath));
+  await writeFile(filepath, content, "utf-8");
+}
+
+export async function fileExists(filepath: string): Promise<boolean> {
+  try {
+    await stat(filepath);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export async function findFiles(dir: string, extension: string = ".md"): Promise<string[]> {
   try {
     const files = await readdir(dir);
-    const filtered = files
-      .filter((file) => file.endsWith(extension))
-      .map((file) => join(dir, file));
-
-    if (ignorePatterns && ignorePatterns.length > 0) {
-      return filterIgnoredFiles(filtered, ignorePatterns);
-    }
-
-    return filtered;
+    return files.filter((file) => file.endsWith(extension)).map((file) => join(dir, file));
   } catch {
     return [];
   }
